refactor: moved provider config to a new private hook
